Divide ios appium_tests to each module
